+++ email_archiver/__init__.py
@@ -1,2 +1,2 @@
 """Email-to-EML Secure Archiver"""
-__version__ = "0.2.0"
+__version__ = "0.3.0"